requirements.txt
aniso8601==8.0.0
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
flask-restx==0.1.1
Flask-Session==0.3.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
jsonschema==3.2.0
jusText==2.2.0
lxml==4.5.0
MarkupSafe==1.1.1
mongoengine==0.19.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.0.0
pymongo==3.10.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
requests-file==1.4.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tinysegmenter==0.3
tldextract==2.2.2
trafilatura==0.3.1
urllib3==1.25.8
Werkzeug==0.16.1
zipp==3.0.0
